modification du logiciel de façon a ce que si on met des valeurs vide dans les choix de remplissage alors il nous énonce que la valeur est vide et annule l'action

+++ Gestionnaire_de_l_Infrastructure_Reseaux/affichage/FenetreSupprimerMateriel.cs
@@ -57,9 +57,11 @@ namespace Gestionnaire_de_l_Infrastructure_Reseaux
 
             if (result == DialogResult.OK)
             {
-                SupprimerElement();
-                MessageBox.Show("élément supprimer de la base de données ");
-                this.Close();
+                if (SupprimerElement())
+                {
+                    MessageBox.Show("élément supprimer de la base de données ");
+                    this.Close();
+                }
             }
             else
             {
@@ -69,26 +71,36 @@ namespace Gestionnaire_de_l_Infrastructure_Reseaux
         }
 
         //supprimer l'élément de la bdd
-        private void SupprimerElement()
+        private bool SupprimerElement()
         {
+            bool flag = false;
             int id;
             var selectedId = comboBox1.SelectedItem as ComboBoxItem;
-            id = selectedId.Id;
-            string query = $"DELETE FROM Materiel_Reseau WHERE id = {id}";
-
-            using (MySqlConnector.MySqlConnection connection = new MySqlConnector.MySqlConnection(comm.connexionBDD()))
+            if (selectedId != null)
             {
-                MySqlConnector.MySqlCommand command = new MySqlConnector.MySqlCommand(query, connection);
-                connection.Open();
-                try
+                id = selectedId.Id;
+                string query = $"DELETE FROM Materiel_Reseau WHERE id = {id}";
+
+                using (MySqlConnector.MySqlConnection connection = new MySqlConnector.MySqlConnection(comm.connexionBDD()))
                 {
-                    command.ExecuteNonQuery();
+                    MySqlConnector.MySqlCommand command = new MySqlConnector.MySqlCommand(query, connection);
+                    connection.Open();
+                    try
+                    {
+                        command.ExecuteNonQuery();
+                    }
+                    catch (Exception ex)
+                    {
+                        MessageBox.Show(ex.Message);
+                    }
                 }
-                catch (Exception ex)
-                {
-                    MessageBox.Show(ex.Message);
-                }
+                flag = true;
             }
+            else
+            {
+                MessageBox.Show("vous n'avez rien choisi", "erreur de saisie");
+            }
+            return flag;
         }
     }
 }
